Updated server.py and client.py
server.py creates a socket server hosted on IP: 0.0.0.0 and Port: 1234, on the local device. If a client connects to the server will display their IP address and port.

client.py is able to connect to the server. Just displays if the client successfully connect to the server or not.