Added the ability for clients to send and receive messages
Added the function 'listen_for_messages' to server.py, which waits for a client to send a message. Then after receiving a message from a client it will send message to all clients on the server using the function 'send_messages_to_all', which goes through all active clients and sends them the message individually using the function 'send_message_to_client'.

Added the function 'comunicate_to_server' to client.py, so that the client can send the username of the client to the server. If the username is not empty client.py sends the username to the client. Then the client will be prompted to send their message using the function 'send_message_to_server'. The client also runs the function 'listen_for_messages_from_server' which will display messages received from the server.